# docs/swagger.yaml
basePath: /api
definitions:
...
  api.FailedResponse:
    properties:
      code:
        type: string
      msg:
        type: string
    type: object
...
  api.MspSuccessResponse:
    properties:
      code:
        type: string
      data:
        items:
          type: string
        type: array
      msg:
        type: string
    type: object
...
  api.TxSuccessResponse:
    properties:
      code:
        type: string
      data:
        $ref: '#/definitions/model.Transaction'
        type: object
      msg:
        type: string
    type: object
  api.TxsSuccessResponse:
    properties:
      code:
        type: string
      data:
        items:
          $ref: '#/definitions/model.Transaction'
        type: array
      msg:
        type: string
    type: object
  api.TxsWeekSuccessResponse:
    properties:
      code:
        type: string
      data:
        items:
          type: integer
        type: array
      msg:
        type: string
    type: object
...
  model.Transaction:
    properties:
      block_num:
        type: integer
      chaincode_function:
        type: string
      chaincode_function_parameters:
        type: string
      chaincode_name:
        type: string
      channel_id:
        type: integer
      create_msp_id:
        type: string
      created_at:
        type: string
      endorser_msp_id:
        type: string
      endorser_signature:
        type: string
      id:
        type: integer
      read_key_list:
        type: string
      read_set:
        type: string
      tx_filter:
        type: integer
      tx_hash:
        type: string
      tx_num:
        type: integer
      tx_response:
        type: integer
      tx_type:
        type: string
      updated_at:
        type: string
      write_key_list:
        type: string
      write_set:
        type: string
    type: object
host: 127.0.0.1:30052
info:
  contact:
    name: API Support
    url: http://www.swagger.io/support
  license:
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Ledgerdata refiner api doc
  version: "1.0"
paths:
...
  /tx/getcreatemspfromtxs/{channel_id}:
    get:
      consumes:
      - application/json
      description: get create MSP list from txs
      parameters:
      - description: channel id
        in: path
        name: channel_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: code":200, "msg:"OK", "data":[msp1,msp2,msp3]}
          schema:
            $ref: '#/definitions/api.MspSuccessResponse'
        "500":
          description: code":3001, "msg:"failed to get create msp list"}
          schema:
            $ref: '#/definitions/api.FailedResponse'
      tags:
      - Transaction
  /tx/gettxbyblocknum:
    post:
      consumes:
      - application/json
      description: get txs by block number
      parameters:
      - description: channel id
        in: path
        name: channel_id
        required: true
        type: integer
      - description: page num
        in: path
        name: page_num
        type: integer
      - description: page count
        in: path
        name: page_count
        type: integer
      - description: sort by(DESC ASC)
        in: path
        name: sort_by
        type: string
      - description: block number
        in: path
        name: block_num
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: code":200, "msg:"OK", "data":[tx1,tx2,tx3]}
          schema:
            $ref: '#/definitions/api.TxsSuccessResponse'
        "500":
          description: code":3001, "msg:"failed to get txs"}
          schema:
            $ref: '#/definitions/api.FailedResponse'
      tags:
      - Transaction
  /tx/gettxbytxhash/{channel_id}/{tx_hash}:
    get:
      consumes:
      - application/json
      description: get tx by tx hash
      parameters:
      - description: channel id
        in: path
        name: channel_id
        required: true
        type: integer
      - description: tx hash
        in: path
        name: tx_hash
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: code":200, "msg:"OK", "data":{tx}}
          schema:
            $ref: '#/definitions/api.TxSuccessResponse'
        "500":
          description: code":3002, "msg:"failed to get tx"}
          schema:
            $ref: '#/definitions/api.FailedResponse'
      tags:
      - Transaction
  /tx/gettxcountbyday:
    post:
      consumes:
      - application/json
      description: get count of transactions of day
      parameters:
      - description: channel id
        in: body
        name: channel_id
        required: true
        schema:
          type: integer
      - description: start time
        in: body
        name: start_time
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: code":200, "msg:"OK", "data":{tx1, tx2}}
          schema:
            $ref: '#/definitions/api.TxsWeekSuccessResponse'
        "500":
          description: code":3002, "msg:"failed to get tx"}
          schema:
            $ref: '#/definitions/api.FailedResponse'
      tags:
      - Transaction
  /tx/gettxcountbymonth:
    post:
      consumes:
      - application/json
      description: get count of transactions of month
      parameters:
      - description: channel id
        in: body
        name: channel_id
        required: true
        schema:
          type: integer
      - description: start time
        in: body
        name: start_time
        required: true
        schema:
          type: string
      - description: end time
        in: body
        name: end_time
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: code":200, "msg:"OK", "data":{tx1, tx2}}
          schema:
            $ref: '#/definitions/api.TxsWeekSuccessResponse'
        "500":
          description: code":3002, "msg:"failed to get tx"}
          schema:
            $ref: '#/definitions/api.FailedResponse'
      tags:
      - Transaction
  /tx/gettxcountbyorg:
    get:
      consumes:
      - application/json
      description: get count of txs of orgs
      produces:
      - application/json
      responses:
        "200":
          description: code":200, "msg:"OK", "data":{tx1, tx2}}
          schema:
            $ref: '#/definitions/api.TxsSuccessResponse'
        "500":
          description: code":3002, "msg:"failed to get tx"}
          schema:
            $ref: '#/definitions/api.FailedResponse'
      tags:
      - Transaction
  /tx/gettxcountbyweek:
    post:
      consumes:
      - application/json
      description: get count of transactions of week
      parameters:
      - description: channel id
        in: body
        name: channel_id
        required: true
        schema:
          type: integer
      - description: start time
        in: body
        name: start_time
        required: true
        schema:
          type: string
      - description: end time
        in: body
        name: end_time
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: code":200, "msg:"OK", "data":{tx1, tx2}}
          schema:
            $ref: '#/definitions/api.TxsWeekSuccessResponse'
        "500":
          description: code":3002, "msg:"failed to get tx"}
          schema:
            $ref: '#/definitions/api.FailedResponse'
      tags:
      - Transaction
  /tx/gettxsbytimerange:
    post:
      consumes:
      - application/json
      description: get transactions by time range
      parameters:
      - description: channel id
        in: body
        name: channel_id
        required: true
        schema:
          type: integer
      - description: sort by
        in: body
        name: sort_by
        schema:
          type: string
      - description: start time
        in: body
        name: start_time
        schema:
          type: string
      - description: end time
        in: body
        name: end_time
        schema:
          type: string
      - description: page number
        in: body
        name: page_num
        schema:
          type: integer
      - description: page count
        in: body
        name: page_count
        schema:
          type: integer
      - description: create msp
        in: body
        name: create_msp
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: code":200, "msg:"OK", "data":{tx1, tx2}}
          schema:
            $ref: '#/definitions/api.TxsSuccessResponse'
        "500":
          description: code":3002, "msg:"failed to get tx"}
          schema:
            $ref: '#/definitions/api.FailedResponse'
      tags:
      - Transaction
swagger: "2.0"